backend/requirements.txt
-e git+https://github.com/pierredelattre/AllocineAPI@f0d6a3e5d26947cb7f3d26104f5fe46d9795d862#egg=allocine_seances
annotated-types==0.7.0
anyio==4.10.0
beautifulsoup4==4.13.5
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
fastapi==0.116.1
geographiclib==2.1
geopy==2.4.1
h11==0.16.0
idna==3.10
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
python-multipart==0.0.20
requests==2.32.5
sniffio==1.3.1
soupsieve==2.8
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
